+++ alpine/src/version_sticker.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-version=$(cat /etc/os-release | grep PRETTY_NAME | sed 's/PRETTY_NAME=//')
-
-echo "$version"