+++ ubuntu/src/startup/version_of.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+### @accetto, October 2019
+
+case "$1" in
+    ubuntu)
+        ### source example: Ubuntu 18.04.3 LTS
+        echo $(cat /etc/os-release 2>/dev/null | grep -Po '(?<=VERSION\=")[0-9.]+')
+        ;;
+esac